Will not -f push branch if there are no changes

+++ actions/create-pull-request/dev/pull-request.ps1
@@ -2,7 +2,8 @@ param (
     [Parameter(Mandatory = $false)]$title = "Automated Pull Request",
     [Parameter(Mandatory = $false)]$branch = "automatic-pull-request",
     [Parameter(Mandatory = $false)]$body = "This is an automated pull request",
-    [Parameter(Mandatory = $false)]$remote = "origin"
+    [Parameter(Mandatory = $false)]$remote = "origin",
+    [Parameter(Mandatory = $false)]$commitMessage = "Create-Pull-Request commit"
 )
 
 $uniqueUserName = "github-actions-automated-pullrequest-$($branch)"
@@ -47,28 +48,32 @@ if ($RemoteString -like "*$branch*") {
     # Branch that is present in remote and has correct last author is expected to have a pull request history.
     $PRstate = (gh pr view $branch)[1]
 }
-else{
+else {
     # If new branch needs to be created, there is no pull request history.
     $PRstate = "NONEXISTENT"
     Write-Host "There is no remote branch named $($branch)"
 }
-# Pushing - set upstream to remote 
-git push --set-upstream $remote $branch -f
 # Adding changes
 git add .
+
+git commit -m $commitMessage
+# Pushing - set upstream to remote 
+$isDiff = (git diff "$($remote)/$($baseBranch)")
+if ($null -eq $isDiff) {
+    write-host "There is no changes. Exiting Create-Pull-Request action.."
+    exit
+}
+git push --set-upstream $remote $branch -f
+# IMORTANT: Writing over current remote branch. The branch will have no commit history.
+git push --force
 
 # What state is the last pull request in?
 if ($PRstate -like "*CLOSED*" -or $PRstate -like "*MERGED*" -or $PRstate -like "*NONEXISTENT*") {
     # if pull request is not active, create new Pull Request
-    git commit -m "Initial Commit Creating Pull Request"
-    # IMORTANT: Writing over current remote branch. The branch will have no commit history.
-    git push --force
     gh pr create --base $baseBranch --head $branch --title $title --body $body
 }
 else {
-    # if pull request is active (open or draft), this action will only push changes. The pull request will automatically update according to new changes.
-    git commit -m "Overwriting Pull Request with $($PRstate)"
-    git push --force
+    # if pull request is active (open or draft). The pull request will try to create new pull request.
     try {
         gh pr create --base $baseBranch --head $branch --title $title --body $body
         write-host "A new pull request was created"
